changed CPI-Config, usability fixes
CPMD format will be tracked through only one variable
-> angstrom and relative coordinates can no longer be combined
-> BUT: now can use format given in GUI
also:
- fixed CPParamWidget not saving changes
- made JSON parsing more lax, don't throw and exit when config is invalid

+++ libvipster/step.h
@@ -81,7 +81,7 @@ public:
         --(this->pse_ptr);
         return *this;
     }
-    AtomListIterator& operator+=(int i){
+    AtomListIterator& operator+=(long i){
         idx += i;
         this->coord_ptr += i;
         this->name_ptr += i;
@@ -89,7 +89,7 @@ public:
         this->pse_ptr += i;
         return *this;
     }
-    AtomListIterator& operator-=(int i){
+    AtomListIterator& operator-=(long i){
         idx -= i;
         this->coord_ptr -= i;
         this->name_ptr -= i;
@@ -97,11 +97,11 @@ public:
         this->pse_ptr -= i;
         return *this;
     }
-    AtomListIterator operator+(int i){
+    AtomListIterator operator+(long i){
         AtomListIterator copy{*this};
         return copy+=i;
     }
-    AtomListIterator operator-(int i){
+    AtomListIterator operator-(long i){
         AtomListIterator copy{*this};
         return copy-=i;
     }
